new changes

+++ Safari Wave/Models/DTOs/Users/CreateUserDTO.cs
@@ -6,7 +6,7 @@
         public string FullName { get; set; } = null!;
         public string Email { get; set; } = null!;
 
-        public int PhoneNo { get; set; }
+        public decimal PhoneNo { get; set; }
 
         public string Address { get; set; } = null!;
